backend/devices/urls.py
"""
设备模块URL配置
"""

from django.urls import path, include
from rest_framework.routers import DefaultRouter
from devices.views import DeviceViewSet

router = DefaultRouter()
router.register('', DeviceViewSet, basename='device')

urlpatterns = [
    path('', include(router.urls)),
]
