requirements.txt
# 数据处理
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# 深度学习
torch>=2.0.0

# 后端
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
channels>=4.0.0
channels-redis>=4.1.0

# 数据库（使用SQLite，无需额外依赖；如需MySQL请安装mysqlclient）
# mysqlclient>=2.2.0

# 模型部署
joblib>=1.3.0
psutil>=5.9.0

# 工具
python-dotenv>=1.0.0
tqdm>=4.66.0
matplotlib>=3.8.0
seaborn>=0.13.0
